requirements.txt
llama-index-core>=0.10.57
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-llms-anthropic>=0.1.0
llama-index-readers-file>=0.1.0
chromadb>=0.5.0
streamlit>=1.32.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.25.0
pytest>=8.0.0
pytest-cov>=5.0.0
